requirements.txt
streamlit
streamlit-drawable-canvas
numpy
pillow
torch
matplotlib
tqdm
gym